Added Cellicium contribution: gregexp: Regular expressions with submatches
This module extends the regexp module in OTP R7B-1 with support for submatches
(the \(...\) syntax in SED regular expressions). This makes it possible to
retrieve several components of a match with a single evaluation of a regexp.

I (Mickaël Rémond) checked that it is still working on OTP R9C-0.

See http://www.cellicium.com/erlang/contribs/ for the initial contribution.